refactor: fixed no-cycle lint error
